Made payload type int in jitter buffers.
Added array of jitter buffers to MprDejitter. They are initialized when we select codecs in MpRtpInputAudioConnection - before 1st rtp frame is received.

+++ sipXmediaLib/include/mp/MpJitterBufferDefault.h
@@ -37,7 +37,7 @@ public:
    //@{
 
    MpJitterBufferDefault(const UtlString& name,
-      uint8_t payloadType,
+      int payloadType,
       unsigned int frameSize);
 
    /**
